requirements.txt
altair
attrs
beautifulsoup4
blinker
cachetools
certifi
cffi
charset-normalizer
clarabel
click
contourpy
curl_cffi
cvxpy
cycler
ecos
exceptiongroup
filelock
fonttools
fpdf
frozendict
fsspec
gitdb
GitPython
greenlet
h11
hf-xet
huggingface-hub
idna
Jinja2
joblib
jsonschema
jsonschema-specifications
kiwisolver
MarkupSafe
matplotlib
mpmath
multitasking
narwhals
networkx
nltk
numpy
osqp
outcome
packaging
pandas
patsy
peewee
pillow
platformdirs
plotly
protobuf
psycopg2-binary
pyarrow
pycparser
pydeck
pyparsing
pyportfolioopt
PySocks
python-dateutil
python-dotenv
pytz
PyYAML
qdldl
referencing
regex
requests
rpds-py
safetensors
scikit-learn
scipy
scs
selenium
six
smmap
sniffio
sortedcontainers
soupsieve
SQLAlchemy
statsmodels
streamlit
sympy
tenacity
threadpoolctl
tokenizers
toml
torch
tornado
tqdm
transformers
trio
trio-websocket
typing_extensions
tzdata
undetected-chromedriver
urllib3
webdriver-manager
websocket-client
websockets
wsproto
xgboost
yfinance